Move test to be actually run by CI
